requirements.txt
hydra-core
numpy
scipy
networkx[default]
matplotlib